fix: throw when setContext is called after an await during SSR
The client already rejects this via `set_context_after_init` (#17031), but
the server had no equivalent check, so `setContext` after a top-level
`await` silently succeeded even though the context may be set too late for
siblings that have already rendered.

Mark the SSR context as initialized once `Renderer.run` schedules its
post-await thunks, and check that flag in `setContext`. Marking in `pop()`
the way the client does would not work, because `Renderer.child` rebuilds
the context with a spread before the component body runs, so `run` captures
a copy that a later mutation never reaches.

`set_context_after_init` moves from client to shared errors so both
runtimes throw the same code.
